* updates to ensure using 64-bit parse options on the command line * many dox updates
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@4740 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ main.cpp
@@ -17,7 +17,7 @@
 
 #include <map>
 
-extern int parse_options;
+extern int64 parse_options;
 extern int warnings, qore_lib_options;
 extern const char *def_charset;
 extern const char *cl_pgm, *exec_class_name, *eval_arg, *cmd_zone;
